changed calculator division to accept doubles

+++ src/main/java/Calculator.java
@@ -5,14 +5,17 @@ public class Calculator {
     }
 
     public int subtract(int number1, int number2){
+
         return number1 - number2;
     }
 
     public int multiply(int number1, int number2){
+
         return number1 * number2;
     }
 
-    public double divide(int number1, int number2){
+    public double divide(double number1, double number2){
+
         return number1 / number2;
     }
